fix(hol-guard): address package shim status review feedback
Extract status enrichment into package_shim_status.py, keep audit fields
nullable for stable contracts, and fix ruff formatting.
