+++ src/codex_plugin_scanner/guard/shims.py
@@ -13,6 +13,7 @@ from pathlib import Path
 from typing import TYPE_CHECKING
 
 from .launcher import merge_guard_launcher_env
+from .package_shim_status import enrich_package_shim_status_payload, record_package_shim_audit_result
 from .shim_probe import (
     SHIM_PROBE_ENV_VALUE,
     SHIM_PROBE_ENV_VAR,
@@ -380,47 +381,6 @@ def activate_package_shims(
     }
 
 
-def _enrich_package_shim_status_payload(
-    status: dict[str, object],
-    manifest: dict[str, object],
-) -> dict[str, object]:
-    bypasses = status.get("bypasses", [])
-    path_broken_managers = sorted(
-        {
-            str(entry["manager"])
-            for entry in bypasses
-            if isinstance(entry, dict) and isinstance(entry.get("manager"), str)
-        }
-    )
-    last_test_at = status.get("last_test_at", {})
-    tested_managers = sorted(
-        manager
-        for manager in last_test_at
-        if isinstance(manager, str) and isinstance(last_test_at, dict)
-    ) if isinstance(last_test_at, dict) else []
-    last_audit_proof_at = manifest.get("last_audit_at")
-    normalized_last_audit = last_audit_proof_at if isinstance(last_audit_proof_at, str) else None
-    normalized_last_tests = last_test_at if isinstance(last_test_at, dict) else {}
-
-    enriched = dict(status)
-    enriched["path_broken_managers"] = path_broken_managers
-    enriched["tested_managers"] = tested_managers
-    enriched["pathBrokenManagers"] = path_broken_managers
-    enriched["testedManagers"] = tested_managers
-    enriched["last_intercept_proof_at"] = normalized_last_tests
-    enriched["lastInterceptProofAt"] = normalized_last_tests
-    enriched["detectedManagers"] = list(status.get("detected_managers", []))
-    enriched["protectedManagers"] = list(status.get("protected_managers", []))
-    enriched["installedManagers"] = list(status.get("installed_managers", []))
-    enriched["activeManagers"] = list(status.get("active_managers", []))
-    enriched["missingManagers"] = list(status.get("missing_managers", []))
-    enriched["undetectedManagers"] = list(status.get("undetected_managers", []))
-    if normalized_last_audit is not None:
-        enriched["last_audit_proof_at"] = normalized_last_audit
-        enriched["lastAuditProofAt"] = normalized_last_audit
-    return enriched
-
-
 def package_shim_status(context: HarnessContext) -> dict[str, object]:
     manifest = _load_package_shim_manifest(context)
     installed_managers = [
@@ -489,27 +449,27 @@ def package_shim_status(context: HarnessContext) -> dict[str, object]:
         path_contains_shim_dir=path_contains_shim_dir,
         shell_profile_configured=bool(profile_status["shell_profile_configured"]),
     )
-    return _enrich_package_shim_status_payload(
+    return enrich_package_shim_status_payload(
         {
-        "active_managers": active_managers,
-        "detected_managers": detected_managers,
-        "installed_managers": installed_managers,
-        "last_test_at": normalized_last_tests,
-        "protected_managers": protected_managers,
-        "path_active": bool(installed_managers) and len(protected_managers) == len(installed_managers),
-        "path_contains_shim_dir": path_contains_shim_dir,
-        "path_status": activation_path_status,
-        "bypasses": bypasses,
-        "manager_details": manager_details,
-        "manifest_path": str(_package_shim_manifest_path(context)),
-        "missing_managers": missing_managers,
-        "restart_shell_required": activation_path_status == "restart_required",
-        "shell_profile_configured": bool(profile_status["shell_profile_configured"]),
-        "shell_profile_path": profile_status["shell_profile_path"],
-        "shell_hints": _path_export_hints(shim_dir),
-        "shim_dir": str(shim_dir),
-        "supported_managers": list(package_shim_supported_managers()),
-        "undetected_managers": undetected_managers,
+            "active_managers": active_managers,
+            "detected_managers": detected_managers,
+            "installed_managers": installed_managers,
+            "last_test_at": normalized_last_tests,
+            "protected_managers": protected_managers,
+            "path_active": bool(installed_managers) and len(protected_managers) == len(installed_managers),
+            "path_contains_shim_dir": path_contains_shim_dir,
+            "path_status": activation_path_status,
+            "bypasses": bypasses,
+            "manager_details": manager_details,
+            "manifest_path": str(_package_shim_manifest_path(context)),
+            "missing_managers": missing_managers,
+            "restart_shell_required": activation_path_status == "restart_required",
+            "shell_profile_configured": bool(profile_status["shell_profile_configured"]),
+            "shell_profile_path": profile_status["shell_profile_path"],
+            "shell_hints": _path_export_hints(shim_dir),
+            "shim_dir": str(shim_dir),
+            "supported_managers": list(package_shim_supported_managers()),
+            "undetected_managers": undetected_managers,
         },
         manifest,
     )
@@ -853,16 +813,6 @@ def _load_package_shim_manifest(context: HarnessContext) -> dict[str, object]:
 
 def _write_package_shim_manifest(context: HarnessContext, payload: dict[str, object]) -> None:
     _package_shim_manifest_path(context).write_text(json.dumps(payload, sort_keys=True), encoding="utf-8")
-
-
-def record_package_shim_audit_result(
-    context: HarnessContext,
-    *,
-    audited_at: str | None = None,
-) -> None:
-    manifest = _load_package_shim_manifest(context)
-    manifest["last_audit_at"] = audited_at or datetime.now(timezone.utc).isoformat()
-    _write_package_shim_manifest(context, manifest)
 
 
 def _record_package_shim_test_results(
